+++ demo/site/server.js
@@ -30,17 +30,7 @@ app.prepare()
                         return;
                     }
                 }
-                if (parsedUrl.pathname === "/access-token-service-worker.js") {
-                    const filename = require.resolve("@comet/cms-site/access-token-service-worker.ejs");
-                    const { mtime } = fs.statSync(filename);
-                    res.setHeader("Cache-Control", "public, max-age=0");
-                    res.setHeader("Content-Type", "application/javascript; charset=UTF-8");
-                    res.setHeader("Last-Modified", mtime.toUTCString());
-                    const str = fs.readFileSync(filename).toString();
-                    res.end(str.replace(/<%= API_URL %>/g, process.env.API_URL));
-                } else {
-                    await handle(req, res, parsedUrl);
-                }
+                await handle(req, res, parsedUrl);
             } catch (err) {
                 console.error("Error occurred handling", req.url, err);
                 res.statusCode = 500;