+++ packages/admin/cms-admin/src/index.ts
@@ -19,7 +19,6 @@ export { BuildRuntime } from "./builds/BuildRuntime";
 export { Publisher } from "./builds/Publisher";
 export { includeInvisibleContentContext } from "./common/apollo/links/includeInvisibleContentContext";
 export { DropdownMenuItem } from "./common/DropdownMenuItem";
-export { AuthorizationErrorPage } from "./common/errors/AuthorizationErrorPage";
 export { BuildInformationProvider } from "./common/header/about/build-information/BuildInformationProvider";
 export { useBuildInformation } from "./common/header/about/build-information/useBuildInformation";
 export { Header } from "./common/header/Header";